Add submit logic to presentation list (WIP)
This does NOT call any endpoints at the moment.
